deploy/_pkgcheck_final/backend/requirements.txt
django==4.2.*
djangorestframework==3.15.*
django-cors-headers==4.3.*
django-filter==24.3
channels==4.0.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
langchain>=0.3.0
langchain-ollama>=0.2.0
requests>=2.31.0
Pillow>=10.0.0
openpyxl>=3.1.0
